requirements.txt
pandas
google-cloud-bigquery
google-api-core